Move default answers and helpers into a lib file
